Tests Just Compiling . [Green]

+++ CampaignEngine.Tests/CampaignScenarios.cs
@@ -1,3 +1,4 @@
+using CampaignEngine.DomainObjects;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -9,7 +10,7 @@ namespace CampaignEngine.Tests
     {
         [Theory(Skip = "Ideation Phase [Red]")]
         [ClassData(typeof(NoCampaignApplicable_TestScenarioData))]
-        public void Given_No_Campaign_Applicable_Then_Engine_Should_return_Net_total(List<Tuple<string , double , int >> cartItems, Double total)
+        public void Given_No_Campaign_Applicable_Then_Engine_Should_return_Net_total(List<Tuple<string , decimal , int >> cartItems, decimal total)
         {
             var shoppingCart = new ShoppingCart();
             cartItems.ForEach((parameters) => shoppingCart.AddItemWithPriceAndQuantity(parameters.Item1, parameters.Item2 , parameters.Item3));
@@ -20,7 +21,7 @@ namespace CampaignEngine.Tests
 
         [Theory(Skip ="Ideation phase [Red]")]
         [ClassData(typeof(NItemsBelongingToCampaign_TestScenarioData))]
-        public void Given_N_Items_Belonging_To_Campaign_Are_Added_Then_Engine_Should_Apply_Discount(List<Tuple<string, double, int>> cartItems, Double total)
+        public void Given_N_Items_Belonging_To_Campaign_Are_Added_Then_Engine_Should_Apply_Discount(List<Tuple<string, decimal, int>> cartItems, decimal total)
         {
             var shoppingCart = new ShoppingCart();
             cartItems.ForEach((parameters) => shoppingCart.AddItemWithPriceAndQuantity(parameters.Item1, parameters.Item2, parameters.Item3));
@@ -31,7 +32,7 @@ namespace CampaignEngine.Tests
 
         [Theory(Skip = "Ideation phase [Red]")]
         [ClassData(typeof(FixedPriceForComboCampaign_TestScenarioData))]
-        public void Given_Sufficient_items_Belonging_To_Fixed_Price_Combo_Campaign_Are_Added_Then_Engine_Should_Apply_Discount(List<Tuple<string, double, int>> cartItems, Double total)
+        public void Given_Sufficient_items_Belonging_To_Fixed_Price_Combo_Campaign_Are_Added_Then_Engine_Should_Apply_Discount(List<Tuple<string, decimal, int>> cartItems, decimal total)
         {
             var shoppingCart = new ShoppingCart();
             cartItems.ForEach((parameters) => shoppingCart.AddItemWithPriceAndQuantity(parameters.Item1, parameters.Item2, parameters.Item3));
