+++ CampaignEngine.Tests/NoCampaignApplicable_TestScenarioData.cs
@@ -10,8 +10,8 @@ namespace CampaignEngine.Tests
         {
             base._data = new List<object[]>
             {
-                new object[] { new List<(string, double, int)> { ("A" , 50 , 1), ("B", 30, 1), ("C", 10, 1) }, 90 },
-                new object[] { new List<(string, double, int)> { ("B" , 40 , 1), ("C", 20, 1), ("D", 15, 1) }, 75 }
+                new object[] { new List<(string, decimal, int)> { ("A" , 50 , 1), ("B", 30, 1), ("C", 10, 1) }, 90 },
+                new object[] { new List<(string, decimal, int)> { ("B" , 40 , 1), ("C", 20, 1), ("D", 15, 1) }, 75 }
             };
         }
